refactor: planned event edit page
